feat: validate TempoApiRequirer relation has limit=1 at instantiation
Previously, TempoApiRequirer.get_data() checked if len(relations)>1 and raised otherwise.  This change moves that validation to the object instantiation, where we require the relation we're wrapping has limit=1.

To do this, we've changed the object's API to accept a charm's RelationMeta object for the wrapped relation instead of simply relation_name.

Although technically unrelated, also changed here is the TempoProviderApi to also accept relation_meta instead of just relation_name.  This change was to keep the API between the two wrappers consistent.