refactor: change TempoCoordinatorCharm._external_url, add ._most_external_url
Previously, TempoCoordinatorCharm._external_url returned the external url if it was available, otherwise it returned the ingress url.  This meant that any consuming code that explicitly wanted the external_url or nothing needed to duplicate checks that exist in external_url, making it easy to later have these go out of sync.

This change breaks _external_url into two functions:
* _external_url, which explicitly returns only an external url or None (if it isn't available)
* _most_external_url, which returns the external_url if it is available, otherwise the internal url

This way the failover behaviour of the previous external_url is preserved, but we can also access exactly the external_url when needed.