refactor: use `_external_url` in `_update_tempo_api_relations`
Previously in `_update_tempo_api_relations`, to determine if the external url we were using was in fact an external url we had a non-intuitive comparison between internal_url and external_url.  Now, we use the explicit `_external_url` property to get exactly the external_url, removing the need for the previous check.